# openapi/model_counted_feature.go
/*
 * PANGEA License Manager
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type CountedFeature struct {

	Name string `json:"name"`

	Version float32 `json:"version"`

	Count int32 `json:"count"`
}