# openapi/model_feature.go
/*
 * PANGEA License Manager
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Feature struct {

	Name string `json:"name"`

	IsPackage bool `json:"isPackage,omitempty"`

	Description string `json:"description,omitempty"`
}