# openapi/model_hardware_key.go
/*
 * PANGEA License Manager
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type HardwareKey struct {

	Id string `json:"id"`

	Kind string `json:"kind"`

	Comments string `json:"comments"`

	CurrentOwnerId int32 `json:"currentOwnerId"`
}