# openapi/model_history_item.go
/*
 * PANGEA License Manager
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type HistoryItem struct {

	OrgName string `json:"orgName"`

	WhenIssued string `json:"whenIssued"`
}