# openapi/model_licensed_feature.go
/*
 * PANGEA License Manager
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type LicensedFeature struct {

	CountedFeature CountedFeature `json:"countedFeature"`

	// YYYY-MM-DD date
	Start string `json:"start"`

	// YYYY-MM-DD date
	End string `json:"end"`
}