# openapi/model_organization.go
/*
 * PANGEA License Manager
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Organization struct {

	Id int32 `json:"id"`

	Name string `json:"name"`

	Contacts string `json:"contacts"`

	Comments string `json:"comments"`
}